+++ dtslint/ts3.5/ReaderEither.ts
@@ -67,3 +67,16 @@ pipe(
   _.bind('a', () => _.of<unknown, string, number>(1)),
   _.bind('b', () => _.of<unknown, string, string>('b'))
 )
+
+//
+// filterOrElseW
+//
+
+// $ExpectType ReaderEither<{ c: boolean; }, "a" | "b", number>
+pipe(
+  _.left<{ c: boolean }, 'a', number>('a'),
+  _.filterOrElseW(
+    (result) => result > 0,
+    () => 'b' as const
+  )
+)